flask_frontend/requirements.txt
# Flask Web Server
flask>=3.0.0
flask-cors>=4.0.0

# Environment Management
python-dotenv>=1.0.1

# Deep Research Dependencies (from parent project)
langgraph>=0.5.4
langchain-community>=0.3.9
langchain-openai>=0.3.28
langchain-anthropic>=0.3.15
langchain-tavily
langchain-google-genai>=2.1.5
tavily-python>=0.5.0
openai>=1.99.2
beautifulsoup4==4.13.3
requests>=2.32.3
markdownify>=0.11.6
